factored parser into lib dir
