+++ lib/parser.js
@@ -1,10 +1,5 @@
 'use strict';
 
-const readline = require('readline');
-const fs = require('fs');
-
-const INFILE = 'resources/extra/Extra_Class_Pool.txt';
-
 const SUBELEMENT_PATTERN = /SUBELEMENT [TGE]\d /;
 const GROUP_PATTERN = /[TGE]\d[A-Z] /;
 const ITEM_PATTERN = /[TGE]\d[A-Z]\d\d \([A-D]\)/;
@@ -12,7 +7,7 @@ const OPTION_PATTERN = /[A-D]\. \w+/;
 const TERMINATION_PATTERN = /~~/;
 const END_PATTERN = /~~~~End of question pool text~~~~/;
 
-var Parser = function() {
+exports.Parser = function() {
   this.item = null;
   this.option = null;
   this.group = null;
@@ -124,17 +119,3 @@ var Option = function(s) {
   this.letter = s.substring(0,1);
   this.text = s.substring(3, s.length);
 }
-
-var p = new Parser();
-
-readline.createInterface({
-  input: fs.createReadStream(INFILE),
-  terminal: false
-}).on('line',
-  function(line) {
-    p.next(line);
-}).on('close',
-	function () {
-    console.log(JSON.stringify(p, null, 4));
-	});
-
